feature: wip design sell page
