+++ src/layout/SideBar.tsx
@@ -2,6 +2,7 @@ import { Drawer, ListItem, ListItemIcon, ListItemText, makeStyles } from '@mater
 import Icon from '@material-ui/core/Icon';
 import React from 'react';
 import { Link } from 'react-router-dom';
+import menu from './menu';
 
 const useStyles = makeStyles(theme => ({
   menu: {
@@ -16,54 +17,6 @@ const useStyles = makeStyles(theme => ({
     marginRight: '1em'
   }
 }));
-
-const menu = [
-  {
-    name: 'Dashboard',
-    to: '/',
-    icon: 'mail'
-  },
-  {
-    name: 'Accounts',
-    to: '/accounts',
-    icon: 'mail'
-  },
-  {
-    name: 'War',
-    to: '/war',
-    icon: 'mail'
-  },
-  {
-    name: 'Auction History',
-    to: '/auction_history',
-    icon: 'mail'
-  },
-  {
-    name: 'Sell',
-    to: '/auction/sell',
-    icon: 'euro_symbol'
-  },
-  {
-    name: 'Story',
-    to: '/story',
-    icon: 'mail'
-  },
-  {
-    name: 'Draw Card',
-    to: '/draw_card',
-    icon: 'mail'
-  },
-  {
-    name: 'Recruit',
-    to: '/recruit',
-    icon: 'mail'
-  },
-  {
-    name: 'Cards',
-    to: '/cards',
-    icon: 'mail'
-  }
-];
 
 const SideBar = () => {
   const classes = useStyles();
