+++ src/layout/menu.ts
@@ -1,0 +1,49 @@
+const menu = [
+  {
+    name: 'Dashboard',
+    to: '/',
+    icon: 'mail'
+  },
+  {
+    name: 'Accounts',
+    to: '/accounts',
+    icon: 'mail'
+  },
+  {
+    name: 'War',
+    to: '/war',
+    icon: 'mail'
+  },
+  {
+    name: 'Auction History',
+    to: '/auction_history',
+    icon: 'mail'
+  },
+  {
+    name: 'Sell',
+    to: '/auction/sell',
+    icon: 'euro_symbol'
+  },
+  {
+    name: 'Story',
+    to: '/story',
+    icon: 'mail'
+  },
+  {
+    name: 'Draw Card',
+    to: '/draw_card',
+    icon: 'mail'
+  },
+  {
+    name: 'Recruit',
+    to: '/recruit',
+    icon: 'mail'
+  },
+  {
+    name: 'Cards',
+    to: '/cards',
+    icon: 'mail'
+  }
+];
+
+export default menu;
